# contracts/openapi/backend-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "VALIDATION_ERROR",
          "SESSION_NOT_FOUND",
          "REVISION_CONFLICT",
          "IDEMPOTENCY_CONFLICT",
          "COMMAND_OUT_OF_ORDER",
          "COMMAND_REJECTED",
          "SNAPSHOT_VERSION_UNSUPPORTED",
          "MANUAL_TICK_DISABLED",
          "SIMULATION_EXECUTION_FAILED",
          "DATABASE_UNAVAILABLE",
          "EVENT_BROKER_UNAVAILABLE",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "EventEnvelope": {
        "properties": {
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "session_revision": {
            "title": "Session Revision",
            "type": "integer"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "tick": {
            "title": "Tick",
            "type": "integer"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "cursor",
          "session_id",
          "session_revision",
          "tick",
          "type"
        ],
        "title": "EventEnvelope",
        "type": "object"
      },
      "EventListResponse": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventEnvelope"
            },
            "title": "Events",
            "type": "array"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "events"
        ],
        "title": "EventListResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "devops-tycoon-backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/game-sessions/{session_id}/events": {
      "get": {
        "operationId": "list_events_api_v1_game_sessions__session_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "description": "Exclusive cursor; returns events with cursor > this value.",
            "in": "query",
            "name": "after_cursor",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Exclusive cursor; returns events with cursor > this value.",
              "minimum": 0,
              "title": "After Cursor",
              "type": "integer"
            }
          },
          {
            "description": "Max events to return. Defaults to 100, capped at 1000. If the result count equals the effective limit, request the next page using the last event's cursor as after_cursor. Cursors are globally monotonic, so a session's cursors ascend but may have gaps (not lost events).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Max events to return. Defaults to 100, capped at 1000. If the result count equals the effective limit, request the next page using the last event's cursor as after_cursor. Cursors are globally monotonic, so a session's cursors ascend but may have gaps (not lost events).",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "SESSION_NOT_FOUND \u2014 unknown session id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "VALIDATION_ERROR \u2014 malformed request (schema, strict ints, unknown command)."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "INTERNAL_ERROR \u2014 unhandled server fault."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "DATABASE_UNAVAILABLE."
          }
        },
        "summary": "List Events",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}